[CO-354] Introduce `NetworkMagic`, routing through `ProtocolMagic`
We add the `NetworkMagic` datatype, which is isomorphic to `Maybe Word8`
but provides more context & type-safety.

For ease of configuration, we also introduce the `RequiresNetworkMagic`
helper datatype, which actually goes into the configuration file,
providing a richer descriptor than a Bool (with functional equivalence).

Unlike our first approach, we route `RequiresNetworkMagic` through the
`ProtocolMagic` datatype, in order to minimize the number of LOCs
changed (limiting our "error surface", as this is a relatively complex
PR).

merged with "[CO-354] WIP: rest of changes for PMPrime"