Simplify the restart-restoration-task interface.
The previous approach (moving the restoration around manually during a
switchToFork) was not working out well, so here is a simplified
approach. A background restoration task can now be cancelled;
switchToFork does this before making the switch. In addition, there is
a "restart" operation that resets the wallet restoration based on the
current tip and utxo, starting the restoration over from the genesis
block.

The "restore" wallet operation was modified so that you can restore an
existing wallet. This operation now atomically deletes the wallet and
re-creates it.